backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
anthropic==0.25.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
transformers==4.41.0
torch==2.3.0
langdetect==1.0.9
deep-translator==1.11.4
fpdf2>=2.7.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiofiles>=23.2.1
colorama==0.4.6
rich==13.7.1
